Enable serial console by default for pxe
If a problem happens in IPA, it can help to use the serial console to
debug. We can also use the serial console to log the output in CI
systems.

We specify 2 console options, one for the serial console, and one for
the graphical display. According to the kernel docs, the last devices
will be used for /dev/console, but output will appear on all of them[1].
`tty0` is last so if there's an emegency prompt needed it'll be put
on the graphical console by default.

[1] https://www.kernel.org/doc/html/v5.5-rc2/admin-guide/serial-console.html